+++ test/lint-fix.test.js
@@ -71,6 +71,31 @@ describe('Lint & fix with API', function () {
 
     }).timeout(200000);
 
+    it('(API:source) should lint and fix (no lintagainafterfix)', async () => {
+        const sampleFilePath = './lib/example/SampleFile.groovy';
+        const prevFileContent = fse.readFileSync(sampleFilePath).toString();
+        const npmGroovyLintConfig = {
+            source: prevFileContent,
+            sourcefilepath: sampleFilePath,
+            fix: true,
+            nolintafter: true,
+            output: 'none',
+            verbose: true
+        };
+        const linter = await new NpmGroovyLint(
+            npmGroovyLintConfig, {
+            jdeployRootPath: 'jdeploy-bundle'
+        }).run();
+
+        assert(linter.status === 0, 'Status is 0');
+        assert(linter.lintResult.summary.totalFixedNumber >= 975, 'Errors have been fixed');
+        assert(linter.lintResult.files[0].updatedSource &&
+            linter.lintResult.files[0].updatedSource !== prevFileContent,
+            'Source has been updated');
+        assert(!linter.outputString.includes('NaN'), 'Results does not contain NaN');
+
+    }).timeout(200000);
+
     it('(API:file) should lint and fix a Jenkinsfile in one shot', async function () {
         const tmpDir = await copyFilesInTmpDir();
         const prevFileContent = fse.readFileSync(tmpDir + '/Jenkinsfile').toString();